Change payout back to instance without initial :( Change IEM cost to integer because choices aren't working.

+++ apps/shows/models.py
@@ -74,7 +74,7 @@ class Show(models.Model):
     commission_check_no = models.IntegerField(blank=True, null=True)
     sound_cost = models.DecimalField(max_digits=10, decimal_places=2, blank=True, null=True)
     sound_check_no = models.IntegerField(blank=True, null=True)
-    in_ears_cost = models.DecimalField(max_digits=10, decimal_places=2, blank=True, null=True, choices=IEM_CHOICES, default=130)
+    in_ears_cost = models.IntegerField(blank=True, null=True, choices=IEM_CHOICES, default=130)
     in_ears_check_no = models.IntegerField(blank=True, null=True)
     costs_itemized = models.BooleanField(default=False)
     print_ship_cost = models.DecimalField(max_digits=10, decimal_places=2, blank=True, null=True)
